requirements.txt
numpy
ahrs
adafruit-circuitpython-gps
#RPi.GPIO
pyserial
smbus2
gpiozero